% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differens_between_outputs.R
\name{differens_between_outputs}
\alias{differens_between_outputs}
\title{Title}
\usage{
differens_between_outputs(
  path = NULL,
  type = "control",
  output_version_old = NULL,
  output_version_new = NULL,
  rfmo = c("CCAMLR", "CECAF", "GFCM", "IATTC", "ICCAT", "ICES", "IOTC", "NAFO", "SEAFO",
    "SPRFMO", "WCPFC", "WECAFC")
)
}
\arguments{
\item{variables}{}
}
\description{
Title
}
