% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_load_cl_landing_rdb_data.R
\name{global_load_cl_landing_rdb_data}
\alias{global_load_cl_landing_rdb_data}
\title{Load CL landing RDB data}
\usage{
global_load_cl_landing_rdb_data(input_path_file_rcg_stats)
}
\arguments{
\item{input_path_file_rcg_stats}{{\link[base]{character}} expected. Input rcg stats data file path (.csv format expected).}
}
\value{
A tibble.
}
\description{
Process for for load the data from CL landing RDB data file.
}
