% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_load_eurostat_data.R
\name{global_load_eurostat_data}
\alias{global_load_eurostat_data}
\title{Load EUROSTAT landings files}
\usage{
global_load_eurostat_data(path)
}
\arguments{
\item{path}{{\link[base]{character}} expected. Input path directory where tsv EUROSTAT landings files are located.}
}
\value{
The function return a {\link[base]{data.frame}}.
}
\description{
Process for Load EUROSTAT landings files (http://ec.europa.eu/eurostat/web/fisheries/data/database).
}
