% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwp_table_2_1.R
\name{rwp_table_2_1}
\alias{rwp_table_2_1}
\title{Regional Work Plan table 2.1}
\usage{
rwp_table_2_1(
  reference_period_start,
  reference_period_end,
  country,
  eurostat,
  rcg_stats,
  fides,
  rfmo,
  path_input,
  path_in_confidential_data,
  path_output
)
}
\arguments{
\item{reference_period_start}{{\link[base]{integer}} expected. Start of reference period. Be careful, the process needs 3 years at least to run.}

\item{reference_period_end}{{\link[base]{integer}} expected. End of reference period. Be careful, the process needs 3 years at least to run.}

\item{country}{{\link[base]{character}} expected. Country of interest. Use 2-alpha country.}

\item{eurostat}{{\link[base]{logical}} expected. Landing statistics downloaded from EUROSTAT (https://ec.europa.eu/eurostat).}

\item{rcg_stats}{{\link[base]{logical}} expected. Landing statistics from the regional database.}

\item{fides}{{\link[base]{character}} expected. {\link[base]{character}} expected. Source of FIDES data. You can choose between "unique" for national extraction from FIDES data (one file per country and years), "common" for global FIDES data for all EU countries or "public" to use a public source (https://griffincarpenter.org/reports/european-fishing-quotas-2001-2021/).}

\item{rfmo}{{\link[base]{character}} expected. RFMO's list to include in output.}

\item{path_input}{{\link[base]{character}} expected. Input path for input files.}

\item{path_in_confidential_data}{{\link[base]{character}} expected. Input path for confidential files.}

\item{path_output}{{\link[base]{character}} expected. Output path for output files.}
}
\value{
I don't know yet dude!
}
\description{
Process for generation the table 2.1, list of required species/stocks, for a Regional Work Plan (RWP).
}
