% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwp_table_2_1_template.R
\name{rwp_table_2_1_template}
\alias{rwp_table_2_1_template}
\title{Regional Work Plan table 2.1 template generation}
\usage{
rwp_table_2_1_template(
  reference_period_start,
  reference_period_end,
  country,
  landing_statistics,
  rfmo,
  input_path,
  output_path = NULL
)
}
\arguments{
\item{reference_period_start}{{\link[base]{integer}} expected. Start of reference period. Be careful, the process needs 3 years at least to run.}

\item{reference_period_end}{{\link[base]{integer}} expected. End of reference period. Be careful, the process needs 3 years at least to run.}

\item{country}{{\link[base]{character}} expected. Country of interest. Use 2-alpha country.}

\item{landing_statistics}{{\link[base]{character}} expected. Landing data statistics source. You can choose between EUROSTAT source (use argument "eurostat", https://ec.europa.eu/eurostat) or regional database source (use argument "rcg_stats").}

\item{rfmo}{{\link[base]{character}} expected. RFMO's list to include in output.}

\item{input_path}{{\link[base]{character}} expected. Input path for input files.}

\item{output_path}{{\link[base]{character}} expected. Output path. By default NULL.}
}
\value{
A list with two elements: "table_2_1_template" and "table_2_1_template_control".
}
\description{
Process for generation the table 2.1 template, list of required species/stocks, for a Regional Work Plan (RWP).
}
