% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwptool-package.R
\docType{package}
\name{rwptool-package}
\alias{rwptool}
\alias{rwptool-package}
\title{rwptool package}
\description{
R toolbox package for the EU Regional Work Plans development.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ob7-ird.github.io/rwptool/}
  \item Report bugs at \url{https://github.com/OB7-IRD/rwptool/issues}
}

}
\author{
\strong{Maintainer}: Mathieu Depetris \email{mathieu.depetris@ird.fr} (\href{https://orcid.org/0000-0001-8080-0531}{ORCID})

Authors:
\itemize{
  \item Joel Vigneau \email{joel.vigneau@ifremer.fr}
  \item Kirsten Birch Håkansson \email{kih@aqua.dtu.dk}
  \item Nuno Prista \email{nuno.prista@slu.se}
}

}
\keyword{internal}
