% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_load_cl_landing_rdbes_data.R
\name{global_load_cl_landing_rdbes_data}
\alias{global_load_cl_landing_rdbes_data}
\title{Load CL landing RDBES data}
\usage{
global_load_cl_landing_rdbes_data(input_path_cl_landing_rdbes_data)
}
\arguments{
\item{input_path_cl_landing_rdbes_data}{{\link[base]{character}} expected. Input CL landing RDBES data file path (.csv format expected).}
}
\value{
A tibble.
}
\description{
Process for for load the data from CL landing RDBES data file.
}
