% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rwp_table_2_1_template.R
\name{rwp_table_2_1_template}
\alias{rwp_table_2_1_template}
\title{Regional Work Plan table 2.1 template generation}
\usage{
rwp_table_2_1_template(
  reference_period_start,
  reference_period_end,
  eu_countries = c("AUT", "BEL", "BGR", "HRV", "CYP", "CZE", "DNK", "EST", "FIN", "FRA",
    "DEU", "GRC", "HUN", "IRL", "ITA", "LVA", "LTU", "LUX", "MLT", "NLD", "POL", "PRT",
    "ROU", "SVK", "SVN", "ESP", "SWE"),
  landing_statistics,
  rfmo = c("CCAMLR", "CECAF", "GFCM", "IATTC", "ICCAT", "ICES", "IOTC", "NAFO", "SEAFO",
    "SPRFMO", "WCPFC", "WECAFC"),
  input_path_directory_eurostat = NULL,
  input_path_file_rcg_stats = NULL,
  input_path_file_fides,
  output_path = NULL
)
}
\arguments{
\item{reference_period_start}{{\link[base]{integer}} expected. Start of reference period. Be careful, the process needs 3 years at least to run.}

\item{reference_period_end}{{\link[base]{integer}} expected. End of reference period. Be careful, the process needs 3 years at least to run.}

\item{eu_countries}{{\link[base]{character}} expected. European Union country(ies) id(s) for data extraction associated. Use 3-alpha country. By default the 27 EU member states.}

\item{landing_statistics}{{\link[base]{character}} expected. Landing data statistics source. You can choose between EUROSTAT source (use argument "eurostat", https://ec.europa.eu/eurostat/web/fisheries/data/database) or regional database source (use argument "rcg_stats").}

\item{rfmo}{{\link[base]{character}} expected. RFMO's list to include in output. By default CCAMLR, CECAF, GFCM, IATTC, ICCAT, ICES, IOTC, NAFO, SEAFO, SPRFMO, WCPFC, WECAFC.}

\item{input_path_directory_eurostat}{{\link[base]{character}} expected. Input path directory for input eurostat files.}

\item{input_path_file_rcg_stats}{{\link[base]{character}} expected. Input path and file name for input CL landing RDB file.}

\item{input_path_file_fides}{{\link[base]{character}} expected. Input path and file name for input FIDES file.}

\item{output_path}{{\link[base]{character}} expected. Output path. By default NULL.}
}
\value{
A list with two elements: "table_2_1_template" and "table_2_1_template_control".
}
\description{
Process for generation the table 2.1 template, list of required species/stocks, for a Regional Work Plan (RWP).
}
